+++ cypress/tests/core/blocks/blocks-image.js
@@ -1,7 +1,9 @@
 describe('Blocks Tests', () => {
   beforeEach(() => {
+    cy.intercept('GET', `/**/*?expand*`).as('content');
+    cy.intercept('GET', '/**/Document').as('schema');
     cy.intercept('POST', '*').as('saveImage');
-    cy.intercept('GET', '/**/image.png/@@images/image').as('getImage');
+    cy.intercept('GET', '/**/image.png/@@images/image-*').as('getImage');
     // given a logged in editor and a page in edit mode
     cy.visit('/');
     cy.autologin();
@@ -11,12 +13,10 @@ describe('Blocks Tests', () => {
       contentTitle: 'My Page',
     });
     cy.visit('/my-page');
-    cy.waitForResourceToLoad('@navigation');
-    cy.waitForResourceToLoad('@breadcrumbs');
-    cy.waitForResourceToLoad('@actions');
-    cy.waitForResourceToLoad('@types');
-    cy.waitForResourceToLoad('my-page');
+    cy.wait('@content');
+
     cy.navigate('/my-page/edit');
+    cy.wait('@schema');
   });
 
   afterEach(() => {
@@ -30,7 +30,7 @@ describe('Blocks Tests', () => {
     cy.get('.ui.basic.icon.button.block-add-button').click();
     cy.get('.ui.basic.icon.button.image').contains('Image').click();
     cy.get('.block.image .ui.input input[type="text"]').type(
-      `https://github.com/plone/volto/raw/master/logos/volto-colorful.png{enter}`,
+      `https://github.com/plone/volto/raw/main/logos/volto-colorful.png{enter}`,
     );
     cy.get('#toolbar-save').click();
     cy.url().should('eq', Cypress.config().baseUrl + '/my-page');
@@ -39,7 +39,7 @@ describe('Blocks Tests', () => {
     cy.get('#page-document img').should(
       'have.attr',
       'src',
-      'https://github.com/plone/volto/raw/master/logos/volto-colorful.png',
+      'https://github.com/plone/volto/raw/main/logos/volto-colorful.png',
     );
 
     cy.get('#page-document img')
@@ -117,7 +117,7 @@ describe('Blocks Tests', () => {
     // then image src must be equal to image name
     cy.get('.block img')
       .should('have.attr', 'src')
-      .and('eq', '/my-page/image.png/@@images/image');
+      .and('contains', '/my-page/image.png/@@images/image-');
 
     cy.get('.block img')
       .should('be.visible')
@@ -146,7 +146,7 @@ describe('Blocks Tests', () => {
 
     cy.get('.block img')
       .should('have.attr', 'src')
-      .and('eq', '/image.png/@@images/image');
+      .and('contains', '/image.png/@@images/image-');
 
     cy.get('.block img')
       .should('be.visible')